+++ src/common/CString.h
@@ -500,7 +500,7 @@ size_t Str_ParseCmds(TCHAR * pCmdLine, INT64 * piCmd, size_t iMax, LPCTSTR pSep 
 * @param iCount max iterations.
 * @return the index of string if success, -1 otherwise.
 */
-int FindTable(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount);
+int FindTable(LPCTSTR pszFind, const LPCTSTR *ppszTable, int iCount);
 /**
 * @brief Look for a string in a table (binary search).
 * @param pszFind string we are looking for.
@@ -508,7 +508,7 @@ int FindTable(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount);
 * @param iCount max iterations.
 * @return the index of string if success, -1 otherwise.
 */
-int FindTableSorted(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount);
+int FindTableSorted(LPCTSTR pszFind, const LPCTSTR *ppszTable, int iCount);
 /**
 * @brief Look for a string header in a table (uses Str_CmpHeadI to compare instead of strcmpi).
 * @param pszFind string we are looking for.
@@ -516,7 +516,7 @@ int FindTableSorted(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount);
 * @param iCount max iterations.
 * @return the index of string if success, -1 otherwise.
 */
-int FindTableHead(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount);
+int FindTableHead(LPCTSTR pszFind, const LPCTSTR *ppszTable, int iCount);
 /**
 * @brief Look for a string header in a table (binary search, uses Str_CmpHeadI to compare instead of strcmpi).
 * @param pszFind string we are looking for.
@@ -525,7 +525,7 @@ int FindTableHead(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount);
 * @param iElemSize size of elements of the table.
 * @return the index of string if success, -1 otherwise.
 */
-int FindTableHeadSorted(LPCTSTR pszFind, const LPCTSTR *ppszTable, size_t iCount, size_t iElemSize = sizeof(LPCTSTR));
+int FindTableHeadSorted(LPCTSTR pszFind, const LPCTSTR *ppszTable, int iCount, size_t iElemSize = sizeof(LPCTSTR));
 
 #define Str_GetTemp static_cast<AbstractSphereThread *>(ThreadHolder::current())->allocateBuffer
